requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
bidict==0.23.1
bottle==0.13.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
propcache==0.3.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
